Remove unneeded additional modules (#1)
