--- .github/workflows/publish-release.yaml
@@ -1,48 +1,0 @@
-name: Publish Release
-on:
-  push:
-    tags:
-      - '*'
-  workflow_dispatch:
-
-jobs:
-  publish:
-    # remove this check when your secrets are setup and you're ready to publish
-    if: ${{ github.repository == 'cortinico/kotlin-android-template'}}
-    runs-on: [ubuntu-latest]
-    env:
-      GRADLE_OPTS: -Dorg.gradle.parallel=false
-
-    steps:
-
-      - name: Checkout Repo
-        uses: actions/checkout@v3
-
-      - name: Setup Java
-        uses: actions/setup-java@v3
-        with:
-          distribution: 'zulu'
-          java-version: '17'
-
-      - name: Setup Gradle
-        uses: gradle/gradle-build-action@v2
-
-      - name: Publish to Maven Local
-        run: ./gradlew publishToMavenLocal
-        env:
-          ORG_GRADLE_PROJECT_SIGNING_KEY: ${{ secrets.ORG_GRADLE_PROJECT_SIGNING_KEY }}
-          ORG_GRADLE_PROJECT_SIGNING_PWD: ${{ secrets.ORG_GRADLE_PROJECT_SIGNING_PWD }}
-
-      - name: Upload Build Artifacts
-        uses: actions/upload-artifact@v3
-        with:
-          name: 'release-artifacts'
-          path: '~/.m2/repository/'
-
-      - name: Publish to the Snapshot Repository
-        run: ./gradlew publishReleasePublicationToNexusRepository
-        env:
-          ORG_GRADLE_PROJECT_SIGNING_KEY: ${{ secrets.ORG_GRADLE_PROJECT_SIGNING_KEY }}
-          ORG_GRADLE_PROJECT_SIGNING_PWD: ${{ secrets.ORG_GRADLE_PROJECT_SIGNING_PWD }}
-          ORG_GRADLE_PROJECT_NEXUS_USERNAME: ${{ secrets.ORG_GRADLE_PROJECT_NEXUS_USERNAME }}
-          ORG_GRADLE_PROJECT_NEXUS_PASSWORD: ${{ secrets.ORG_GRADLE_PROJECT_NEXUS_PASSWORD }}
