+++ app/src/main/res/values/dimens.xml
@@ -1,5 +1,3 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <dimen name="base_grid_size">8dp</dimen>
-    <dimen name="doub_grid_size">16dp</dimen>
 </resources>
