+++ app/src/main/res/values/strings.xml
@@ -1,10 +1,3 @@
 <resources>
-    <string name="app_name">Kotlin Android Template</string>
-    <string name="title">This is just a template</string>
-    <string name="description">You can compute a factorial using the library-kotlin module. The result will be also shown in a notification using the library-android module.</string>
-    <string name="factorial_hint">Insert a number to compute his factorial</string>
-    <string name="compute">Compute</string>
-    <string name="notification_title">The Result is: %s</string>
-    <string name="please_enter_a_number">Please Enter a Number</string>
-    <string name="compose_version">In Compose</string>
+    <string name="app_name">Quies</string>
 </resources>
