+++ buildSrc/src/main/java/Coordinates.kt
@@ -2,16 +2,6 @@ const val PUBLISHING_GROUP = "com.ncorti.kotlin.template"
 
 object AppCoordinates {
     const val APP_ID = "com.ncorti.kotlin.template.app"
-
     const val APP_VERSION_NAME = "1.0.0"
     const val APP_VERSION_CODE = 1
-}
-
-object LibraryAndroidCoordinates {
-    const val LIBRARY_VERSION = "1.0.0"
-    const val LIBRARY_VERSION_CODE = 1
-}
-
-object LibraryKotlinCoordinates {
-    const val LIBRARY_VERSION = "1.0.0"
 }
