--- library-compose/src/main/AndroidManifest.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <application>
-        <activity
-            android:name="com.ncorti.kotlin.template.app.ComposeActivity"
-            android:exported="false" />
-    </application>
-
-</manifest>
