--- library-kotlin/build.gradle.kts
@@ -1,23 +1,0 @@
-version = LibraryKotlinCoordinates.LIBRARY_VERSION
-
-plugins {
-    id("java-library")
-    kotlin("jvm")
-    id("maven-publish")
-    publish
-}
-
-dependencies {
-    testImplementation(libs.junit)
-}
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_17
-    targetCompatibility = JavaVersion.VERSION_17
-    withSourcesJar()
-    withJavadocJar()
-}
-
-kotlin {
-    jvmToolchain(17)
-}
