+++ settings.gradle.kts
@@ -19,7 +19,4 @@ enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 
 include(
     "app",
-    "library-android",
-    "library-compose",
-    "library-kotlin"
 )
